[LZ] Host, Shared and Device allocs with long type
Support allocations using a long value as numBytes.

+++ drivers/spirv/src/main/java/uk/ac/manchester/tornado/drivers/spirv/SPIRVLevelZeroContext.java
@@ -189,7 +189,7 @@ public class SPIRVLevelZeroContext extends SPIRVContext {
                     levelZeroContext.getDefaultContextPtr(), //
                     deviceMemAllocDesc, //
                     hostMemAllocDesc, //
-                    (int) numBytes, //
+                    numBytes, //
                     TornadoOptions.LEVEL_ZERO_BUFFER_ALIGNMENT, //
                     l0Device.getDeviceHandlerPtr(), //
                     deviceBuffer);
@@ -199,7 +199,7 @@ public class SPIRVLevelZeroContext extends SPIRVContext {
             int result = levelZeroContext.zeMemAllocDevice(//
                     levelZeroContext.getDefaultContextPtr(), //
                     deviceMemAllocDesc, //
-                    (int) numBytes, //
+                    numBytes, //
                     TornadoOptions.LEVEL_ZERO_BUFFER_ALIGNMENT, //
                     l0Device.getDeviceHandlerPtr(), //
                     deviceBuffer);
